remove bindActionCreators and fixup some issues
