+++ shared/actions/json/favorite.json
@@ -1,6 +1,6 @@
 {
   "prelude": [
-    "import * as Constants from '../constants/favorite'",
+    "import * as Types from '../constants/types/favorite'",
     "import * as Folders from '../constants/folders'"
   ],
   "actions": {
@@ -22,7 +22,7 @@
     },
     "favoriteList": {},
     "favoriteListed": {
-      "folders": "Constants.FolderState"
+      "folders": "Types.FolderState"
     },
     "favoriteSwitchTab": {
       "showingPrivate": "boolean"
@@ -31,7 +31,7 @@
       "isPrivate": "boolean"
     },
     "kbfsStatusUpdated": {
-      "status": "Constants.KBFSStatus"
+      "status": "Types.KBFSStatus"
     },
     "markTLFCreated": {
       "folder": "Folders.Folder"
