+++ shared/constants/favorite.js
@@ -1,8 +1,8 @@
 // @flow
-import * as RPCTypes from '../constants/types/flow-types'
+import * as RPCTypes from './types/flow-types'
+import * as Types from './types/favorite'
 import {defaultKBFSPath} from './config'
 import {parseFolderNameToUsers, sortUserList} from '../util/kbfs'
-import type {Exact} from '../constants/types/more'
 import type {Folder, MetaType, FolderRPCWithMeta} from './folders'
 import type {UserList} from '../common-adapters/usernames'
 
@@ -11,53 +11,7 @@ export const ExitCodeFuseKextError = 4
 // See KBDefines.h: KBExitFuseKextPermissionError
 export const ExitCodeFuseKextPermissionError = 5
 
-type ListState = any
-// TODO this is super messy and there some entangled flow error. let's revisit this soon
-/* {
-  tlfs?: Array<Folder>,
-  ignored?: Array<Folder>,
-  isPublic: boolean,
-  style?: any,
-  smallMode?: boolean,
-  onClick?: (path: string) => void,
-  onRekey?: (path: string) => void,
-  onOpen?: (path: string) => void,
-  onChat?: (tlf: string) => void,
-  onToggleShowIgnored?: ?() => void,
-  showIgnored?: boolean,
-  extraRows?: Array<React.Node>,
-} */
-
-export type FolderState = {
-  privateBadge: number,
-  private: ListState,
-  publicBadge: number,
-  public: ListState,
-}
-
-export type ViewState = Exact<{
-  showingPrivate: boolean,
-  publicIgnoredOpen: boolean,
-  privateIgnoredOpen: boolean,
-}>
-
-export type KBFSStatus = {
-  isAsyncWriteHappening: boolean,
-}
-
-export type State = Exact<{
-  folderState: FolderState,
-  fuseInstalling: boolean,
-  fuseStatus: ?RPCTypes.FuseStatus,
-  fuseStatusLoading: boolean,
-  kbfsInstalling: boolean,
-  kbfsOpening: boolean,
-  kbfsStatus: KBFSStatus,
-  kextPermissionError: boolean,
-  viewState: ViewState,
-}>
-
-const initialState: State = {
+const initialState: Types.State = {
   folderState: {
     private: {
       isPublic: false,
@@ -193,8 +147,6 @@ function folderFromPath(username: string, path: string): ?Folder {
     return folderFromFolderRPC(username, folderRPC)
   }
 }
-
-export type {Folder as FolderRPC} from '../constants/types/flow-types'
 
 export {
   canonicalizeTLF,
