+++ shared/login/signup/passphrase/index.js
@@ -1,9 +1,8 @@
 // @flow
+import * as Creators from '../../../actions/signup'
 import React, {Component} from 'react'
 import Render, {type Props} from './index.render'
-import {bindActionCreators} from 'redux'
-import {checkPassphrase, restartSignup} from '../../../actions/signup'
-import {connect} from 'react-redux'
+import {connect, type TypedState} from '../../../util/container'
 
 type State = {
   pass1: string,
@@ -16,16 +15,11 @@ type ContainerProps = {
   restartSignup: () => void,
 }
 
+// TODO recompose
 class PassphraseForm extends Component<ContainerProps, State> {
-  state: State
-
-  constructor(props: ContainerProps) {
-    super(props)
-
-    this.state = {
-      pass1: '',
-      pass2: '',
-    }
+  state = {
+    pass1: '',
+    pass2: '',
   }
 
   render() {
@@ -43,9 +37,12 @@ class PassphraseForm extends Component<ContainerProps, State> {
   }
 }
 
-export default connect(
-  state => ({
-    passphraseError: state.signup.passphraseError,
-  }),
-  dispatch => bindActionCreators({checkPassphrase, restartSignup}, dispatch)
-)(PassphraseForm)
+const mapStateToProps = (state: TypedState) => ({
+  passphraseError: state.signup.passphraseError,
+})
+
+const mapDispatchToProps = (dispatch: Dispatch) => ({
+  checkPassphrase: (pass1: string, pass2: string) => dispatch(Creators.checkPassphrase(pass1, pass2)),
+  restartSignup: () => dispatch(Creators.restartSignup()),
+})
+export default connect(mapStateToProps, mapDispatchToProps)(PassphraseForm)
